# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Eidolon API
  version: 1.0.0
  description: |
    REST API for managing AI companies, agents, projects, tasks, and runtime
    operations. All company-scoped endpoints require an authenticated user
    with membership in the requested company.
servers:
  - url: /
    description: Current deployment
...
security:
  - clerkCookie: []
paths:
...
  /api/companies/{companyId}/privacy/subjects/{userId}/export:
    parameters:
      - $ref: '#/components/parameters/CompanyId'
      - $ref: '#/components/parameters/SubjectUserId'
    get:
      tags: [Privacy]
      summary: Export the personal data held for a subject
      description: |
        Owner only (`privacy.manage`). Returns every row in this company that
        references the subject, table by table. Authentication material (MFA
        factors and session rows) is deliberately excluded.

        The subject must already be a member of the named company; any other
        subject returns 404 so an owner cannot probe membership elsewhere.
        Responses are sent with `Cache-Control: no-store`.
      responses:
        '200':
          description: Subject data export
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/SubjectExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller is not an owner of this company
        '404':
          description: No such subject in this company
  /api/companies/{companyId}/privacy/subjects/{userId}/erase:
    parameters:
      - $ref: '#/components/parameters/CompanyId'
      - $ref: '#/components/parameters/SubjectUserId'
    post:
      tags: [Privacy]
      summary: Erase the personal data held for a subject
      description: |
        Owner only (`privacy.manage`). Irreversible. Deletes rows that exist only
        to describe the subject, nulls attribution on company work product, and
        pseudonymises audit and NOT NULL references.

        Check `remainingReferences` in the response: any non-zero value means the
        erasure was incomplete. Identity itself lives in Clerk and must be deleted
        there as well.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmSubject]
              properties:
                confirmSubject:
                  type: string
                  description: Must equal the `userId` path parameter.
                email:
                  type: string
                  format: email
                  description: |
                    Only needed for a pending invitation, which is keyed by email
                    and has no user id to match on.
      responses:
        '200':
          description: Erasure report
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/ErasureReport'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller is not an owner of this company
        '404':
          description: No such subject in this company
        '409':
          description: An owner cannot erase their own subject data
components:
  securitySchemes:
    clerkCookie:
      type: apiKey
      in: cookie
      name: __session
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SubjectUserId:
      name: userId
      in: path
      required: true
      description: External identity provider user id.
      schema:
        type: string
  responses:
    Unauthorized:
      description: Authentication is required
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SubjectExport:
      type: object
      required: [subject, companyId, generatedAt, tables, totalRows]
      properties:
        subject:
          type: string
        companyId:
          type: string
          format: uuid
        generatedAt:
          type: string
          format: date-time
        totalRows:
          type: integer
        tables:
          type: array
          items:
            type: object
            required: [table, column, rows]
            properties:
              table:
                type: string
              column:
                type: string
                description: The column in that table holding the subject id.
              rows:
                type: array
                items:
                  type: object
                  additionalProperties: true
    ErasureReport:
      type: object
      required: [companyId, pseudonym, erasedAt, actions, rowsAffected, remainingReferences]
      properties:
        companyId:
          type: string
          format: uuid
        pseudonym:
          type: string
          description: Irreversible per-company identifier that replaced the subject id.
        erasedAt:
          type: string
          format: date-time
        rowsAffected:
          type: integer
        remainingReferences:
          type: integer
          description: |
            Rows still referencing the subject after erasure, re-counted from the
            database. Any value other than 0 means the erasure was incomplete.
        actions:
          type: array
          items:
            type: object
            required: [table, column, strategy, rowsAffected]
            properties:
              table:
                type: string
              column:
                type: string
              strategy:
                type: string
                enum: [delete, 'null', pseudonymise]
              rowsAffected:
                type: integer
    ErrorResponse:
      type: object
      required: [status, code, message]
      properties:
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        details:
          description: Present for validation errors.